+++ src/is_array.ts
@@ -1,0 +1,3 @@
+export function isArray<T>(value: any): value is Array<T> {
+  return value instanceof Array;
+}